Add ability to retrieve source via HTTP
The app is now able to retrieve both local files with source code and
remote HTTP sites such as github. Both HTTP and HTTPS are supported
using the erlang apps ssl and inets. The two apps are automically
started using the command line in rel/files/vm.args.

The current approach is very simplistic and can crash the dig server
if the HTTP response is not ok. This commit is a proof of concept
that files and HTTP can be mixed.

The previous attempt to use github.com/yrashk/evfs was removed since
there is a conflict between evfs and yaws. Loading evfs as a standalone
app or as an included app causes yaws to crash on startup.

Examples of valid file paths for to load:

  /Users/jay/Gitdk/etdd/dig_and_delve/src/dig_and_delve.app.src
  /Users/jay/Gitdk/etdd/dig_and_delve/src/
  https://raw.github.com/duomark/etdd/master/etdd_dig/src/etdd_dig_app.erl

Files can be loaded singly by name, or all files ending in ".erl" from
a given directory. For now, URLs must refer to a specific source file
ending in ".erl" or ".app.src".